requirements.txt
asgiref
certifi
click
colorama
crispy-bootstrap5
distlib
dj-database-url
Django
django-crispy-forms
django-environ
django-filter
djangorestframework
filelock
git-filter-repo
gunicorn
h11
Markdown
MouseInfo
mysql-connector-python
numpy
opencv-contrib-python
packaging
pillow
pipenv
platformdirs
psycopg2
psycopg2-binary
PyAutoGUI
PyGetWindow
PyMsgBox
pyperclip
PyRect
PyScreeze
pytweening
setuptools
sqlparse
typing_extensions
tzdata
uvicorn
virtualenv
whitenoise